refactor: :recycle: 5.5 Blog list frontend, step5
Changed ToggableButton and BlogForm to use ref. In the current configuration, it is not wrong, but it feels awkward because when you call BlogForm, one parameter is missing (setShowToggableButton). It does not matter because if it is not defined, you can manage the error or even verify that it is defined.  The problem with this is that it is not natural... I spent a half an hour trying to understand how it worked after returning from vacation (didn't notice the "cloneElement", which adds this function to the ToggableButton).